Add support for HandleBarsTemplateFromImport
This might never be used since it appears that Safari and
Firefix when polyfilling, flatten the <template> tags into
the base document - which means querySelector() can just find
them.  But since this works - I will keep it in in case,
the polyfill behavior is different in older browsers.

This can probably be deleted with enough experience
with older browsers.  Sigh.  Web Components.