Major refactor of I18N
- Move the util code from U:: to I18N::
- Make the default to use Symfony's translation unless
  $CFG->legacytranslation = true;

These changes should be far less problematic to get working because
the Symfony libraries do not depend on the underlying operating
system's locale support.  There may be a small performance impact
if the gettext() was doing a lot of caching of the mo files.  But
we could built that as well.